feat : zoom 수정
