+++ Assets/_Scripts/Yu/Zoom.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c269ddbf542d5446a53328a8632c1c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
